+++ lib/crm/helper.rb
@@ -1,5 +1,9 @@
 module Crm
   module Helper
-    include Crm::Helpers::Attributes
+    def self.included(base)
+      base.include Crm::Helpers::Attributes
+      base.include Crm::Helpers::Finders
+      base.include Crm::Helpers::Persistence
+    end
   end
 end