+++ changelog.md
@@ -1,5 +1,6 @@
 ### 1.1.1
 
+- Include `Crm::Helpers::Finders` and `Crm::Helpers::Persistence` automagically when including `Crm::Helper`.
 - Rename `assign_attributes` to `assign_crm_attributes`, deprecating `assign_attributes` in the near future.
 
 ### 1.1.0